requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
jinja2==3.1.6
pydantic==2.11.2
openai==1.75.0
python-dotenv==1.1.0
python-telegram-bot[job-queue]==22.5
